+++ packages/gov-ui-kit/src/components/radioGroup/radio/index.ts
@@ -1,0 +1,1 @@
+export { Radio, type IRadioProps } from './radio';